Added more quote stuff

+++ ChewieBot/ChewieBot/Services/Implementation/QuoteService.cs
@@ -5,6 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using ChewieBot.Database.Model;
 using ChewieBot.Database.Repository;
+using ChewieBot.Events;
 
 namespace ChewieBot.Services.Implementation
 {
@@ -13,10 +14,21 @@ namespace ChewieBot.Services.Implementation
         private IQuoteData quoteData;
         private IUserService userService;
 
+        public event EventHandler<QuoteAddedEventArgs> QuoteAddedEvent;
+        public event EventHandler<QuoteDeletedEventArgs> QuoteDeletedEvent;
+
+        private List<Quote> quoteList;
+
         public QuoteService(IQuoteData quoteData, IUserService userService)
         {
             this.quoteData = quoteData;
             this.userService = userService;
+            this.quoteList = this.quoteData.GetAllQuotes().ToList();
+        }
+
+        public IEnumerable<Quote> GetAllQuotes()
+        {
+            return this.quoteList;
         }
 
         public Quote AddQuote(string username, string quoteText)
@@ -24,14 +36,26 @@ namespace ChewieBot.Services.Implementation
             var user = this.userService.GetUser(username);
             if (user != null)
             {
-                return quoteData.AddQuote(user, quoteText);
+                var quote = quoteData.AddQuote(user, quoteText);
+                if (quote != null)
+                {
+                    this.quoteList.Add(quote);
+                    this.QuoteAddedEvent?.Invoke(this, new QuoteAddedEventArgs { Quote = quote, QuoteList = this.quoteList });
+                    return quote;
+                }
             }
             return null;
         }
 
         public void DeleteQuote(int id)
-        {
-            quoteData.DeleteQuote(id);
+        {            
+            var quoteToDelete = this.quoteList.FirstOrDefault(x => x.Id == id);
+            if (quoteToDelete != null)
+            {
+                this.quoteList.Remove(quoteToDelete);
+                this.QuoteDeletedEvent?.Invoke(this, new QuoteDeletedEventArgs { Id = id });
+                quoteData.DeleteQuote(id);
+            }            
         }
 
         public Quote GetQuote(int id)
